requirements.txt
flask
pandas
numpy
scikit-surprise
gunicorn
